Fix for #3045. Fixing  WebException.Status on Unix
This commit fixes #3045.
WebExecption class currently looks at the value of underlying HResult to
set the value of the Status.
We need to set the value of HttpRequestMessage.HResult to appropriate
CURLcode in case of errors. WebException maps this to the relevant
WebExceptionStatus
