+++ src/System.Net.Requests/src/System/Net/WebException.cs
@@ -7,7 +7,7 @@ using System.Net.Http;
 
 namespace System.Net
 {
-    public class WebException : InvalidOperationException
+    public partial class WebException : InvalidOperationException
     {
         private const WebExceptionStatus DefaultStatus = WebExceptionStatus.UnknownError;
 
@@ -62,30 +62,6 @@ namespace System.Net
             {
                 return _response;
             }
-        }
-
-        internal static WebExceptionStatus GetStatusFromException(HttpRequestException ex)
-        {
-            WebExceptionStatus status;
-
-            // Issue 2384: update WebException.GetStatusFromException after System.Net.Http API changes
-            //
-            // For now, we use the .HResult of the exception to help us map to a suitable
-            // WebExceptionStatus enum value.  The .HResult is set into this exception by
-            // the underlying .NET Core and .NET Native versions of the System.Net.Http stack.
-            // In the future, the HttpRequestException will have its own .Status property that is
-            // an enum type that is more compatible directly with the WebExceptionStatus enum.
-            switch (ex.HResult)
-            {
-                case Interop.WININET_E_NAME_NOT_RESOLVED:
-                    status = WebExceptionStatus.NameResolutionFailure;
-                    break;
-                default:
-                    status = WebExceptionStatus.UnknownError;
-                    break;
-            }
-
-            return status;
         }
 
         internal static Exception CreateCompatibleException(Exception exception)
